+++ docs/changelog.md
@@ -10,3 +10,12 @@
 
 * added support for Dbaas Postgres
 * restructured documentation
+
+## 6.1.1
+
+### Enhancements:
+
+* added backup_location to dbaas postgres cluster create
+* added direction, since and until arguments to logs get command
+* added private argument to k8s cluster create
+* added gateway_ip argument to nodepool create